requirements.txt
pyaudio
SpeechRecognition
google-generativeai
gtts
streamlit
langchain
pypdf
sentence-transformers
langchain-community
faiss-cpu